revert usePlusButton copy to ReportWelcomeText component
